web/requirements.txt
coverage==4.0.3
pytz==2015.7
Django==1.9.2
django-braces==1.8.1
django-debug-toolbar==1.4
django-extensions==1.6.1
django-filter==0.12.0
django-jenkins==0.18.1
django-shotgun==0.2
djangorestframework==3.3.3
django-rest-swagger==0.3.5
drf-extensions==0.2.8
drf-hal-json==0.9.0
drf-nested-fields==0.9.3
factory-boy==2.6.1
fake-factory==0.5.3
Markdown==2.6.5
psycopg2==2.6.1
pyslack-real==0.6.0
python-dateutil==2.5.0
requests==2.9.1
six==1.10.0
sphinx-me==0.3
sqlparse==0.1.18
tqdm==3.8.0
urllib3==1.14
wheel==0.29.0
